+++ frontend/app/src/store/statistics/state.ts
@@ -1,0 +1,13 @@
+import { NetValue } from '@/services/types-api';
+import { StatisticsState } from '@/store/statistics/types';
+
+const defaultNetValue: () => NetValue = () => ({
+  times: [],
+  data: []
+});
+
+export const defaultState = (): StatisticsState => ({
+  netValue: defaultNetValue()
+});
+
+export const state: StatisticsState = defaultState();